fix: change order of Union types for media properties
The order of Union types for media properties in various files was
reversed to `Union[str, InputFile]` for consistency. The changes were
made in the following files:
- aiogram/types/input_media_video.py
- aiogram/types/input_media_audio.py
- aiogram/types/input_media_document.py
- aiogram/types/input_media_animation.py
- aiogram/types/input_media_photo.py
- aiogram/utils/media_group.py

fix: update type check in base.py for TelegramObject (dev-3)

Replaced BotContextController with TelegramObject in the type check to
ensure proper handling of non-standard objects.

test: add TODO comments in test files (dev-3)

Added TODO comments to highlight parts of the code that need future
improvements in the following test files:
- tests/test_api/test_types/test_input_file.py
- tests/test_api/test_client/test_session/test